requirements.txt
asgiref==3.3.1
Django==3.1.4
djangorestframework==3.12.2
numpy==1.19.3
pandas==1.1.5
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
